cast some floating point calculation to ULONG in a mui object creation taglist. When compiled under AOS it seems to otherwise end up as float on the stack.
git-svn-id: https://svn.aros.org/svn/aros/trunk@23774 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/system/ftmanager/ftmanager.c
@@ -928,7 +928,7 @@ ULONG fiNew(Class *cl, Object *o, struct opSet *msg)
 			MUIA_String_Accept, "0123456789",
 			MUIA_String_AdvanceOnCR, TRUE,
 			MUIA_String_MaxLen, 6,
-			MUIA_String_Integer, face->max_advance_width * 250.0 / 72.307,
+			MUIA_String_Integer, (ULONG)(face->max_advance_width * 250.0 / 72.307),
 			MUIA_CycleChain, TRUE,
 			End,
 		Child, Label1("Fixed width"),
